[receiver/sqlserver]update resources attributes to export identifying attributes (#39449)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

We used to export those attributes on the log level, this is not
alignihg with the semantic convention, also, the old way will cause
waste of disk resources, as we can group them to a higher level to avoid
repeat them all the places.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->