[extension/oauth2clientauthextension] Allow exporting of OAuth extension to re-use it in other components, required for #39248 (#39514)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
As discussed in issue #39248 - generic OAuth support for kafka exporter
(to be extended/applied for receiver also).
This is only for exporting the OAuth class so we can re-use it. Only
internal references changed as it was not imported up to now. Created as
a separate PR (suggestion by @atoulme in PR #39482).

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Required for #39482 and #39248

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Nothing to be changed, only existing test.

<!--Describe the documentation added.-->
#### Documentation
Nothing to be changed.

<!--Please delete paragraphs that you did not use before submitting.-->