[receiver/k8sobjectsreceiver] Refactor Validate to only validate static config  (#39469)
#### Description

Refactor Validate() to only validate static config without external
calls to Kubernetes API server as discussed in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38851#discussion_r2046811250

Once this is merged, the PR linked above should introduce `error_mode`
implemented in `start` too.

#### Link to tracking issue
It is related to the issue and follow-up PR mentioned above, but it
doesn't fix the issue.

#### Testing

I've refactored the tests in config_test.go, so we are not testing
object discovery and added new tests for additional failure modes.
I've build a dev images, and validated that functionality is the same as
before the refactor.

#### Documentation

This is a "breaking change" in a sense we are changing what we are
validating in Validate. The new validation logic will not fail when
objects are missing. The receiver will fail the same way in this case,
but in `start`.