receiver/kafkareceiver: fix hyphenated text encodings (#39806)
#### Description

Fix support for text encodings with hyphens in their names.

If the encoding name has a hyphen then it is an invalid extension ID,
but we should not return an error due to this if it's a built-in
encoding.

#### Link to tracking issue

Fixes #39793

#### Testing

Added a new unit test covering hyphenated text encoding names (fails
without the associated fix).

#### Documentation

N/A