[receiver/sqlserver] Properly parse scientific notation numbers to ints (#39905)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
SQL Server stores large integers in scientific e notation.
`strconv.ParseInt` fails to parse this and raises an error, as seen in
the original filed issue. For our use case, we still want these to be
considered valid integers, so before failing entirely, attempt to parse
to float and convert to integer.

This change also uses already existing methods, `retrieveInt` and
`retrieveFloat` to be able to re-use the parsing functionality. The
original issue was only for `sqlserver.database.tempdb.space`, but I
believe this issue is relevant for all metrics of type `int`.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #39124

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Changed test values of a couple integers to ensure new functionality
works. The changed values broke tests on `main` in the same way the
filed bug shows, but tests are passing with this change.