Add slash to libhoney paths (#40071)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Adds a slash to the end of the `traces_url_paths:` so that libhoney
appending datasets will always work regardless of whether the user adds
the slash or not.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #40070

<!--Describe what testing was performed and which tests were added.-->
#### Testing

curling a simple json object into the libhoney receiver to see that it
doesn't return a 404 error.

<!--Describe the documentation added.-->
#### Documentation

No changes to the documentation because it is handled in the
configuration parser now.

<!--Please delete paragraphs that you did not use before submitting.-->