[exporter/awss3exporter] Add the retry mode, max attempts and max backoff settings (#39509)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
The queue and timeout settings have been added in #36264. However, there
is no retry settings added. Initially I thought we should rely on the
standard `configretry.BackOffConfig`, and then realized that the AWS S3
client has a "standard" retryer implementation enabled by default. It
also handles the definition of retryable errors
https://pkg.go.dev/github.com/aws/aws-sdk-go-v2/aws/retry#hdr-Standard,
which makes more sense to leverage the AWS SDK instead of shoehorning it
to the standard `configretry.BackOffConfig`.

Explicitly mention the default retry mode in the documentation to avoid
confusion. Also expose the retry settings so the user can have control.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Mentioned in #36264.

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation
Updated the exporter documentation to include the retry settings.

Signed-off-by: Mengnan Gong <namco1992@gmail.com>