[opampsupervisor] log connection message (#40333)
#### Description

Change the `onConnect` log to be an info. This makes is clear when
reading the opampsupervisor logs when the connection happened
successfully. This log should not spam, as the connection should only
happen once when the agent comes online or under other interesting
situations, like needing to reconnect after the server disappeared.