[exporter/prometheusremotewrite] : WAL read success/failure telemetry (#40272)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR introduces below metrics for WAL:

- `otelcol_exporter_prometheusremotewrite_wal_reads`: Number of WAL
reads
- `otelcol_exporter_prometheusremotewrite_wal_reads_failures`: Number of
WAL reads that failed

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Part of #39556 

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added a UT `TestWALRead_Telemetry`. During WAL startup it tries to read
from the WAL but it fails because there is nothing to read. For that
reason both metrics have failed

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Arthur Silva Sens <arthursens2005@gmail.com>