Use consistent commit for update-otel (#40462)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Right now we use 'main' as the commit when updating:

https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/e0f779ce273e3f5de18369e9cb18147d9bdb0d6c/Makefile#L5-L6

This is troublesome if there are pushes to main while the update-otel is
running, or just in general since we go via the Go proxy and sometimes
the latest commit has not been chached. This changes things so that we
use a fixed commit from the checkout.