[collectdreceiver] ensure HTTP server exits before Shutdown returns (#40498)
#### Description

If Shutdown is called immediately after Start, then it can happen that
the HTTP server's Shutdown method returns before the Server method is
called. In this case, the listener may remain active, and causes test
failures like this:

```
=== RUN   TestComponentLifecycle
=== RUN   TestComponentLifecycle/metrics-lifecycle
    generated_component_test.go:67: 
                Error Trace:    /home/andrew/projects/opentelemetry-collector-contrib/receiver/collectdreceiver/generated_component_test.go:67
                Error:          Received unexpected error:
                                listen tcp 127.0.0.1:8081: bind: address already in use
                Test:           TestComponentLifecycle/metrics-lifecycle
--- FAIL: TestComponentLifecycle (0.00s)
    --- FAIL: TestComponentLifecycle/metrics-lifecycle (0.00s)
```

To address that, add a WaitGroup to ensure the Serve method is called
and returns before the receiver's Shutdown method returns.

#### Link to tracking issue

Fixes #40406

#### Testing

Run the unit tests. They were reliably failing for me before, now they
do not.

#### Documentation

N/A