kafkareceiver: Refactor sarama receiver (#40629)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Refactors the sarama receiver and splits the current receiver into more
files to pave the way for a Franz Go client implementation opt-in.

- consumer_message.go: Contains the generic kafkaMessage interface and a
wrapper for Sarama messages.
- consumer_sarama.go: Contains the sarama specific bits for consumer
group and message processing.
- kafka_receiver.go: Keeps the generic logic of the receiver that would
be applicable to both Sarama and Franz Go clients.

I've got a working Franz Go client implementation with superior
performance to Sarama, but I want to keep the PRs small and manageable.

~The performance is 5-7% better when auto-commit is enabled, and
500+-700% better when message marking is enabled. I will post the full
benchmark results in that PR as it's more applicable.~

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Part of #40628

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests all pass.

<!--Describe the documentation added.-->
#### Documentation

N/A

---------

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>
Co-authored-by: Sean Marciniak <30928402+MovieStoreGuy@users.noreply.github.com>