[internal/aws] Mock time.Ticker for deterministic unit tests (#40737)
#### Description
This PR fixes a flaky unit test that was depending on functionality to
complete within 1ms in a unit tests. I've made this test deterministic
by allowing a mocked `time.Ticker` to be used during unit tests, while
preserving the original functionality outside of the unit test
environment.

---

I don't believe a changelog is needed since this is just fixing a flaky
unit test

#### Link to tracking issue
Fixes #27892

Signed-off-by: Michael Shen <mishen@umich.edu>