[chore] [exporter/awscloudwatchlogs] move @boostchicken and @rapphil to emeritus codeowners (#39471)
@boostchicken and @rapphil haven't been responding to issues and pull
requests for a couple months now. See for example
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/30773,
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/31382,
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/37297,
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/39339.

@boostchicken, @rapphil, first of all thank you for all the work you've
done! Please respond here or in CNCF Slack
[#otel-collector-dev](https://cloud-native.slack.com/archives/C07CCCMRXBK)
if you want to keep being active codeowners. You are also welcome to
respond and confirm being moved to emeritus.

Also marks the exporter as unmaintained and seeking new codeowners.

According to the docs on
[codeowners](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.124.1/CONTRIBUTING.md#emeritus-roles),
this pull request should wait for 2 months before being merged.