receiver/prometheusremotewrite: Correctly handle metrics without type metadata (#41005)
#### Description

While trying the component end-to-end, I've noticed we would hit
nilpointer references almost immediately. While accessing this code
path:


https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/377e2ff1dedf4ea14664602c152d4093e767bafb/receiver/prometheusremotewritereceiver/receiver.go#L371

`metric` was nil

The `metric` variable is generated in this code path:


https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/377e2ff1dedf4ea14664602c152d4093e767bafb/receiver/prometheusremotewritereceiver/receiver.go#L340-L367

Suppose it's present in the cache—perfect. It was already assigned from
the start. If it was not present in the cache, we create it in the next
if conditional, but only for the types we support! With this PR, we're
now skipping all metrics without type metadata.

Signed-off-by: Arthur Silva Sens <arthursens2005@gmail.com>