[exporter/elasticsearch] Profiling: Fix storing of sample attributes (#40963)
#### Description
This PR was meant to add the sample attribute `process.executable.name`.
During testing, it became obvious that the iteration over the sample
attributes had a bug.
This is fixed in this PR as well.

#### Link to tracking issue
Fixes #40961

#### Testing
Added a test for reading sample attribute `process.executable.name`.

#### Before
In Universsal Profiling flamegraph, the Executable was always empty and
the Thread name was empty or wrong.

#### After

![Screenshot_20250627_122048](https://github.com/user-attachments/assets/85cae009-9e7c-47a9-a61f-08217fe2dbb5)