[pkg/ottl] Fix OTTL functions by using setters (#39416)
#### Description

Some OTTL functions do not call the accessor.setter function, but
instead change the variable directly.
This doesn't work with the profiles signal where non-trivial processing
is needed.

For example, attributes are split into a lookup table and arrays of
indices (the rationale is de-duplication). While the accessor.getter
function returns a pcommon.Map for attributes, any changes made to that
map require passing the map to the accessor.setter function.

#### Link to tracking issue
Fixes #39100

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->