Performance improvements for the metricstarttime processor (#40998)
#### Description

The existing implementation copied everything from the previous
resourcemetrics to a completely new resourcemetrics. However, the
component is marked as mutating data, so this PR just changes it to
mutate the batch of metrics in-place. This avoids a lot of copying.

From previous experience caching start times and values, storing
attributes and exemplars in a cache is very expensive unless you need
them. The second commit adds "minimal copy to" methods to avoid storing
attributes and exemplars.

#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/39400

#### Testing

Passes all existing unit tests.