kafkaexporter: Improve internal metrics telemetry consistency, measure latency in seconds (#41217)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
In
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/40834
additional metrics were introduced to add internal telemetry emitted by
the exporter. All the duration and latency metrics were recorded in
milliseconds which is inconsistent with other components which record it
in seconds instead. This PR deprecates the metrics that measure the
latency in milliseconds in favor of new set metrics that use seconds
instead. Additionally the changes were made to use more consistent
naming for `messages` -> `records`.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #41108 

<!--Describe what testing was performed and which tests were added.-->
#### Testing
The unit tests were updated to reflect the changes.

<!--Describe the documentation added.-->
#### Documentation
The documentation was regenerated with `mdatagen`.

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Andrew Wilkins <axwalk@gmail.com>