docs: fix receiver example containing invalid encoding (#40956)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This is just a small docs fix removing the optional (and invalid)
`encoding: raw_json` setting. Copy and pasting this example into an otel
config will result in a startup error.
As it is not required I opted to remove it from the minimal example.