chore: Add more unit tests to GCP encoding (#41446)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR relates to the goal of the issue
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41087.

We need to update the encoding to support semantic conventions, but to
make it more clear what the breaking changes will be, we need to update
the unit tests to cover a wider range and check all the resulting fields
of a log.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Relates to #41087.

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Unit tests added.

<!--Describe the documentation added.-->
#### Documentation

N/A.

<!--Please delete paragraphs that you did not use before submitting.-->