Added setting to config to redact data of all types (#37631)
Description
The redaction is done based on its AsString value
rather than SDR. This is based on this issue:

https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/36684

The bug is that credit card numbers (as in the example on the README)
are often provided as int numbers. The current redaction processor
ignores all fields that are not of type int.

Link to tracking issue

https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/36684

Fixes No way to redact non string values

Documentation
I added the setting and turned it to true in the README as redacting
un-delimited credit cards would definitely be under the scope of this
setting.
(This issue didn't pass the CLI due to the first commit not having an
email)

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: opentelemetrybot <107717825+opentelemetrybot@users.noreply.github.com>
Co-authored-by: Yang Song <songy23@users.noreply.github.com>
Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>