[chore] [receiver/hostmetrics] Update documentation to remove Mac exclusion for `disk` and `cpu` (#41513)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Previously, the `disk` and `cpu` scrapers would error on `darwin`
runtime due to the metrics not being supported without CGO. This changed
in a `gopsutil` update to [use `purego` for loading Mac shared objects
instead of CGO bindings](https://github.com/shirou/gopsutil/pull/1702).
These now work as expected on `darwin`, and nothing restricts these
receivers from working.

This PR removes the declaration in the documentation that these scrapers
are unsupported on Mac.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #33393 

<!--Describe what testing was performed and which tests were added.-->
#### Testing
The tests aren't disabled for `darwin` as far as I can tell, so unless
I'm mistaken the tests for these scrapers have been running anyway.

<!--Describe the documentation added.-->
#### Documentation
This is a documentation PR. :smiley: 

<!--Please delete paragraphs that you did not use before submitting.-->