ci: add workflow task to auto-update the JMX scraper (#41551)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Extends the automation of auto-updating the JMX gatherer to also update
the JMX scraper. A PR like
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/41533
should be created.


Renamed the CI file and duplicated the `check` and `update` tasks to
target the Scraper. I opt for duplicating the tasks instead of
parameterize them, due to the JMX Gatherer being in the deprecation
process.

---------

Co-authored-by: Alex Boten <223565+codeboten@users.noreply.github.com>