fix npe on some metrics for vcenter vms (#41701)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

I believe that this should be the only nullable field improperly
accessed within this file.

```go
// The ConfigInfo data object type encapsulates the configuration settings and
// virtual hardware for a virtual machine.
//
// This type holds all the information
// that is present in the .vmx configuration file for the virtual machine.
type VirtualMachineConfigInfo struct {
```

I'm thinking users may have created a machine and attached the vmdk to
it and thats may be the potential cause?

Either way this should resolve the NPE and shouldn't crash the collector
when we encounter these.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #41018

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

Changelog