[chore][exporter/kafka]Fix incorrect queue batch setting for metrics and traces (#41824)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
PR
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/41776
introduced a bug where incorrect queue batch settings were used for
metrics and traces. This would result in panic due to the incorrect
encoding defined in the incorrect queue batch settings when the
`sending_queue#batch` is used with sizers that require the encoding (ex:
bytes sizer). The PR updates the test cases to catch this too.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Related to
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41775

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Updated test cases.

<!--Describe the documentation added.-->
#### Documentation

Not required.

<!--Please delete paragraphs that you did not use before submitting.-->