[tailsamplingprocessor] Add BlockOnOverFlow option (#41546)
#### Description

When running the tailsampling processor as part of a durable collector
pipeline, it's better to block trace intake than to drop old traces.
This option makes it configurable to do so (default false).

#### Testing
Unit tests show the behavior of the new traceLimiter. I've also tested
this change in our infrastructure at Grafana Labs and the pipeline
backed up as expected until the decision timeout. This makes autoscaling
of the TSP simpler to reason about, since we can confidently set an
autoscaling target on in memory traces, and trust that during an
ingestion spike we will slow down until the autoscaler can kick in.