[receiver/windowseventlog]: If remote.domain is configured, pass it into EvtRPCLogin (#41947)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Fixing what appears to be a bug/miss from the remote windows event log
collection. It appears to myself that we're passing down from
windowseventlogreceiver to the pkg/stanza but the actual syscall for
`evtOpenSession` does not actually pass the `domain` field.

Please let me know if I'm missing anything!

Docs for context of the structures being updated to pass the value into.


https://learn.microsoft.com/en-us/windows/win32/api/winevt/ns-winevt-evt_rpc_login


https://learn.microsoft.com/en-us/windows/win32/api/winevt/nf-winevt-evtopensession

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #41950

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added small unit test to ensure we are passing the value using the
existing test framework.

<!--Describe the documentation added.-->
#### Documentation

Changelog