[chore] Fix uses of testify Eventually (#42262)
#### Description

The latest version of testify (v1.11.0) makes it so that
`assert.Eventually(condition, timeout, period)` performs its first check
of `condition` immediately in stead of waiting for the end of the first
`period`. This revealed a few incorrect uses of `Eventually` in the
codebase which make the test fail if the condition is checked too early.