[chore] Fix test flakiness in `prometheusexporter` (#42351)
With the use `t.Context()` there is a race in which the HTTP server
doesn't shutdown on first pass it hits context canceled error since
`t.Cleanup` always cancels the context prior to the clean-up code.
Moving the shutdown to a `defer` ensures that the shutdown happens
before `t.Context()` is cancelled.

Fix #42259