[chore][receiver/prometheus] Add integration test for Kong (#42741)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This PR introduces a new functional test for Prometheus data coming from
a [Kong Gateway](https://developer.konghq.com/). This will help ensure
that the Prometheus receiver can connect to a Kong Gateway, successfully
scrape metrics, and the metrics that are scraped are the proper format
and naming.