[processor/unroll]: upstream unroll processor (#42500)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Adds a new log processor component that will take a log body of slices
and expand them into multiple log entries within the ScopeLogs.

This is a component we made in the [Bindplane Distro of the
Collector](https://github.com/observIQ/bindplane-otel-collector/tree/main).
We initially solved it for logs without proper parsing extensions. This
pairs quite well with the
[transformprocessor](https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/processor/transformprocessor)
to
[Split](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/ab184d2b57c41b4b66e16eac1b53a1856dc72161/pkg/ottl/ottlfuncs/README.md#split)
and unroll say based off improperly formatted newlines depending on the
receiver sending the data.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #42491

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Believe to be properly unit tested, however its been a second since I've
added a component so happy to add some more tests where they are
applicable.

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->