[receiver/azureeventhubreceiver] use `$Default` as default consumer group (#43051)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
The old SDK sets `$Default` as the consumer group when no consumer
groups are set. This is a fix to the new SDK implementation to do the
same. This would cause issues when users attempted to use the new SDK
and didn't have the consumer group set.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #43049 

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added tests for the function for getting the consumer group


<!--Please delete paragraphs that you did not use before submitting.-->