[receiver/k8sobjectsreceiver] Switch to standby mode when leader lease is lost (#43054)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Similar to what we did in #42330, this PR ensures similarity and puts
the receiver in stand-by instead of shutdown when `k8sleaderelector` is
used.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #42706

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Tested locally, and tests were not changed to ensure the current
receiver is not affected at all.

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Signed-off-by: Paulo Dias <paulodias.gm@gmail.com>