[receiver/kafkareceiver] try to fix flaky tests for franz-go consumer (#43083)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Tentative fix for the flaky CI tests on TestConsumerShutdownConsuming.
The failures were showing up as `expected=4, actual=3`, which seems tied
to timing in slower CI runners. This PR adds a small wait to ensure the
expected number of messages is processed before shutting down the
consumer.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #43022

---------

Signed-off-by: Paulo Dias <paulodias.gm@gmail.com>