[chore]Prevent conflict on available port in integrationtest (#43121)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
`testutil::GetAvailablePort` returns available port to be used in test.
In our integration test we require 2 ports, 1 for the OTLP receiver and
the other for mock Elasticsearch. Both these ports are active when the
collector starts, however, the ports need to be found before the
collector starts making the `testutil::GetAvailablePort` flaky as it can
return the same port in consecutive calls. The PR starts a temporary
listener for the sender port when creating the mock ES port to avoid
this conflict.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41353

<!--Describe what testing was performed and which tests were added.-->
#### Testing
N/A

<!--Describe the documentation added.-->
#### Documentation
N/A

<!--Please delete paragraphs that you did not use before submitting.-->