[chore]Remove flaky test TestLRUSetLifeTime (#43171)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

`TestLRUSetLifeTime` is a flaky test, and the root cause seems to be
that the expire feature uses the wall clock to remove items. Prior
attempts to fix this test haven't been successful. Taking a step back
and looking into the test, it seems we are testing the `lifetime`
feature of the go-freelru library more than our own usage of the
library. Creating this PR to drop the test, as I feel it is out of scope
of the exporter. Happy to discuss if others feel this is a required
test.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/43153

<!--Describe what testing was performed and which tests were added.-->
#### Testing
N/A

<!--Describe the documentation added.-->
#### Documentation
N/A

<!--Please delete paragraphs that you did not use before submitting.-->