[exporter/prometheus] Add sending_queue configuration (#42629)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Adds `sending_queue` configuration. 

Since Prometheus is a pull-based exporter I am not sure to what extent
most options on the configuration are useful, please be critical about
that!

The reason I am adding it is so that the batching configuration can be
used: if we move away from the batch processor, users are going to
expect batching here since we recommend batching.