Standardize changelog component list (#43215)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This PR re-applies #43179 with an added fix in the changelog pipeline.

This PR adds some code to generate the list of allowed components for
`chloggen`.
All tooling was already prepared previously so we just need to make use
of the features.
The PR also adds a CI check to validate the component label in all
changelog entries.

<!-- Issue number if applicable -->
#### Link to tracking issue
Fixes #43178
Re-applies #43179

<!--Describe the documentation added.-->
#### Documentation

I added a readme file as a replacement for the code comments in
`.chloggen/config.yaml` since those are removed during auto generation
of the file.

<!--Please delete paragraphs that you did not use before submitting.-->