[chore][sumologicextension]: fix flaky test (#43319)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
A test was flaky due to depending on timing and an exact count of a
periodic condition. By asserting instead that a minimum count was
achieved instead, the problem can be sidestepped.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #42867

<!--Describe what testing was performed and which tests were added.-->
#### Testing
`go test -count 100 -run
TestStoreCredentials_PreexistingCredentialsAreUsed` will fail before
this change is applied, and succeed after.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>